+++ src/components/fields/groupObservations/groupObservations.test.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { CssBaseline, ThemeProvider } from '@mui/material';
+import theme from '../../../services/theme/theme';
+import GroupObservations from './groupObservations';
+import { StoreProvider } from '@ska-telescope/ska-gui-local-storage';
+import { THEME, viewPort } from '../../../utils/testing/cypress';
+
+const value = 0;
+
+describe('<GroupObservations />', () => {
+  for (const theTheme of THEME) {
+    it(`Theme ${theTheme}: Renders`, () => {
+      viewPort();
+      cy.mount(
+        <StoreProvider>
+          <ThemeProvider theme={theme(theTheme)}>
+            <CssBaseline />
+            <GroupObservations value={value} obsId="1" />
+          </ThemeProvider>
+        </StoreProvider>
+      );
+    });
+  }
+});