+++ src/utils/present.tsx
@@ -5,6 +5,14 @@ import 'katex/dist/katex.min.css';
 
 export const presentLatex = (inStr: string) => <Latex>{inStr}</Latex>;
 
+export const presentSensCalcError = (inArr: string, length = 0) => {
+  if (!inArr || inArr.length === 0) {
+    return '';
+  }
+  const arr = inArr.split('\n');
+  return arr[length];
+};
+
 export const presentUnits = (inUnits: string) => {
   switch (inUnits) {
     case 'arcsec2':